Ward API

+++ src/main/java/com/recyclanka/waste_management/repository/MunicipalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MunicipalRepository extends JpaRepository<Municipal, Integer> {
+public interface MunicipalRepository extends JpaRepository<Municipal, Long> {
 }
